retake_tasks/project/the_tube/requirements.txt
astroid==1.5.3
Django==1.11.5
djangorestframework==3.6.4
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
six==1.11.0
wrapt==1.10.11
